config: set up docker with the web app
