--- client/templates/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <script src="{{url_for('static', filename='index.js')}}" defer></script>
-    <h1>TEMP HTML</h1>
-    <button id="btn-record">Record</button>
-    <input type="text" id="input-output-language" placeholder="Output language">
-</body>
-</html>
