+++ docker-compose.yaml
@@ -13,3 +13,10 @@ services:
       - DB_URL=mongodb://172.17.0.1:27017
     depends_on:
       - db
+
+  web-app:
+    build: ./web-app
+    ports:
+    - "5001:5001"
+    depends_on:
+      - translator
